Hopefully improved the build system configuration
Now defines three functions: make, make-clean, and make-serial that are bound
to f7, f8, and S-f7 respectively. Can be redefined for more complicated build
settings (such as a remote build host).

Also added a couple modes to the not-code-modes list (compilation-mode and
gud-mode), and added an attempt at making the vertical window dividing bar
prettier in text mode (but this appears to be broken at the moment).