Better num-compile-threads defaults.
Added system-cores.el, a library written specifically for this by an awesome
SO user, and used it to set num-compile-threads by default to the number of
threads + 1.

Also added a new function add-plugin-dir, which will add the parameter and all
its immediate subdirectories to load-path. This is now being used for
site-lisp.